server/requirements.txt
asgiref==3.5.2
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.1
psycopg2-binary==2.9.4
PyJWT==2.5.0
pyotp==2.7.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
